Reorganized code
