+++ vpc_deployment/tests/unit/test_env_stack.py
@@ -1,15 +1,13 @@
 import aws_cdk as core
 import aws_cdk.assertions as assertions
-
-
-from modules.pipeline.stacks.env_stack import ENVStack
+from modules.vpc_deployment_stack import VPCDeploymentStack
 
 # Some tests to show cdk can be tested with, we could add more tests but intrest of time ,I am adding only few tests.
 
 
 def test_aws_vpc_created():
     app = core.App()
-    stack = ENVStack(app, "vpc-stack")
+    stack = VPCDeploymentStack(app, "vpc-stack")
     template = assertions.Template.from_stack(stack)
     print(template.to_json())
 
@@ -18,7 +16,7 @@ def test_aws_vpc_created():
 
 def test_aws_vpc_created_with_enabled_dns_support():
     app = core.App()
-    stack = ENVStack(app, "vpc-stack")
+    stack = VPCDeploymentStack(app, "vpc-stack")
     template = assertions.Template.from_stack(stack)
 
 
@@ -27,7 +25,7 @@ def test_aws_vpc_created_with_enabled_dns_support():
 
 def test_aws_vpc_created_with_cidr_block():
     app = core.App()
-    stack = ENVStack(app, "vpc-stack")
+    stack = VPCDeploymentStack(app, "vpc-stack")
     template = assertions.Template.from_stack(stack)
